moved database startup into server side
