+++ lib/routing.js
@@ -1,5 +1,5 @@
 import { FlowRouter } from 'meteor/ostrio:flow-router-extra';
-import '../client/pages/device-list.html';
+import '../client/pages/patient-list.html';
 import '../client/main.html';
 
 // main page
@@ -11,9 +11,9 @@ FlowRouter.route("/", {
 });
 
 //device list page
-FlowRouter.route("/device-list", {
-    name: "device-list",
+FlowRouter.route("/patient-list", {
+    name: "patient-list",
     action() {
-        this.render("deviceList");
+        this.render("patientList");
     }
 });
